W226: the typed-indexing cluster — +4 MATCH (6565 -> 6569), nd -1584, one window cleared
Fourth main-model template-transfer wave. 0 MATCHes lost, 0 new over-window, 0 regressions kept.

**Typed struct/array indexing is now the campaign's most productive idiom, and it clusters.**
Replacing hand-computed `*(T*)((u8*)base + i*STRIDE + OFF)` with indexing a real array of a real
struct has now matched six functions across three files. Two of this wave's four came from it back
to back: `mdlFile 0033c0a0` nd58 -> **0** (a long-lived `u32* model` with `model[0xe]` and
`model[0x1c]` replacing chained byte-offset accesses), and then `00342990` nd60 -> **0** by
transferring that *same shape from the function that had just matched*. Each solved function is a
better template than the census can offer, because it is the same author writing the same routine
in the same file. That was broadcast mid-wave and is now standing guidance: **when a fix lands,
try the nearest two or three functions immediately rather than continuing down the list.**

The other two matches:

  itfMesManager 003a5570  nd222 -> **0**, in three sibling-shaped steps: removing a variadic
                          call-cast macro so the callee is called directly (222 -> 197), declaring
                          `param_4` as `u32` to reproduce retail's unsigned byte-to-float
                          conversion (197 -> 113), and staging the `DAT_00959eec` slot before the
                          branch rather than inside it (113 -> 0).
  mt_evtCustomEvent 0039d8e0  nd140 -> **0**.  The result of a `u64`-aliased call was being stored
                          as `u8`, dropping retail's sign-extend-then-zero-extend pair. Declaring
                          it `s8` and converting explicitly to `u8` at the lookup call reproduced
                          it exactly.

**Largest gains, and one window cleared:**

  mt_evtMain 00367a20     nd317 -> **32**, four ordered padded `RwV4d` stack aggregates plus a
                          direct context-based state store.
  _h_camp_subb 0013c240   nd572 -> **420** and **24 bytes over window -> 520 bytes under**, by
                          reusing the file's existing populate and category-clear helpers instead
                          of recomposing field addresses inline. Its neighbour `0013c780` then took
                          the same cluster transfer, nd759 -> **513**.
  mdlFile 00348da0        nd185 -> **22**.
  datCalc 00309080        nd190 -> **44**: two if-chains converted to switches copied from the
                          adjacent `00309250`, the dispatch key `char` -> `u8`, and `s16` casts at
                          three lookup calls.
  btlFormation 002c3ce0   nd173 -> 106; `k_fldFrame 001aaad0` nd106 -> 85.

**Restraint worth recording.** Three agents independently found targets that looked like typed-index
candidates and declined them because no matched sibling supplied the real layout — a 0xe-stride
persona table, a 0x2c battle record, a Camp list-entry work block. **Guessing a stride compiles
silently and misrepresents the data**, so leaving the byte arithmetic in place was the right call
each time. One agent also correctly reported that a census heading I gave it did not exist rather
than inventing a substitute; the list was regenerated from the checkpoint instead.

Also settled: at `btlCamera 002a79f0` replacing a bare `0x00697898` with the named symbol regressed
nd334 -> 666, so that site genuinely requires the bare-address emission shape — the named-symbol
correction that was worth 337 points elsewhere in the same file is per site, like every other
addressing decision in this codebase.

+++ src/Event/mt_evtCustomEvent.c
@@ -4809,14 +4809,14 @@ void FUN_0039d880(int param_1)
 }
 #define FUN_0039d880(...) ((void (*)(...))FUN_0039d880)(__VA_ARGS__)
 #undef FUN_0039d8e0
-// FUN_0039D8E0 NONMATCHING
+// FUN_0039D8E0
 
 
 u32 FUN_0039d8e0(int param_1)
 {
   u32 *config;
   int *piVar1;
-  u8 uVar2;
+  s8 uVar2;
   short sVar3;
   int iVar4;
   u32 uVar5;
@@ -4860,7 +4860,7 @@ u32 FUN_0039d8e0(int param_1)
     break;
   case 2:
     uVar2 = FUN_0016df30_u64((short)piVar1[3]);
-    uVar7 = FUN_0030bbb0(uVar2);
+    uVar7 = FUN_0030bbb0((u8)uVar2);
     FUN_0039ef30(piVar8,0,uVar7);
     uVar7 = FUN_00173220((u16)piVar1[0xd]);
     FUN_0039ef30(piVar8,1,uVar7);
